rename SignUp page directory
